Request new keep with stake lock duration
Keep factory requires to provide a stake lock duration for the newly
opened keep. The value is used to lock signer KEEP token stakes for the
period of keep activity. It is a protection so the tokens cannot be
undelegated before the keep is closed.

Here we pass deposit term as the value for the lock duration.